added getDeeplinkIntent to handler for easier testing
